chore(wallets): extract AddWallet into its own component
