+++ extension/src/popup/components/account/AddWallet/styles.scss
@@ -1,0 +1,52 @@
+@use "../../../styles/utils.scss" as *;
+
+.AddWallet {
+  display: flex;
+  flex-direction: column;
+
+  &__row {
+    background-color: var(--sds-clr-gray-03);
+    border-radius: 6px;
+    margin-bottom: #{pxToRem(12px)};
+    padding: #{pxToRem(12px)} #{pxToRem(16px)};
+    cursor: pointer;
+
+    &__icon {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      width: 30px;
+      height: 30px;
+      border-radius: 50%;
+      border-width: 1px;
+      padding: #{pxToRem(8px)};
+      margin-bottom: #{pxToRem(24px)};
+
+      &.lime {
+        background-color: var(--sds-clr-lime-06);
+        border-color: var(--sds-clr-lime-09);
+      }
+
+      &.purple {
+        background-color: var(--sds-clr-pink-06);
+        border-color: var(--sds-clr-pink-09);
+      }
+
+      &.blue {
+        background-color: var(--sds-clr-navy-06);
+        border-color: var(--sds-clr-navy-09);
+      }
+    }
+
+    &__title {
+      color: var(--sds-clr-white);
+    }
+
+    &__description {
+      color: var(--sds-clr-gray-11);
+      font-weight: var(--font-weight-medium);
+      margin-bottom: #{pxToRem(8px)};
+      margin-top: #{pxToRem(6px)};
+    }
+  }
+}
